feature/bugfix: fixed redirects of shopify onboarding flow from user install.
